requirements.txt
# --- Unit 2: Data Manipulation & Analysis ---
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# --- Unit 2: Data Visualization ---
matplotlib>=3.7.0
seaborn>=0.12.0

# --- Unit 3 & 4: Machine Learning & Deep Learning ---
scikit-learn>=1.2.0
tensorflow>=2.12.0
joblib>=1.3.0

# --- Unit 5: Computer Vision & Satellite Data ---
opencv-python>=4.7.0
earthengine-api>=0.1.350  # Essential for connecting to Satellite Data API
geemap>=0.20.0            # For interactive mapping with Earth Engine
folium>=0.14.0            # To render maps in your app
Pillow>=9.0.0

# --- Unit 6: Generative AI ---
# (Uses TensorFlow/Keras for GANs as per your curriculum)

# --- Unit 7: Model Deployment ---
streamlit>=1.22.0
python-dotenv>=1.0.0      # For managing API keys securely
folium
streamlit-folium  #Streamlit + maps = folium is required
geemap   #geemap = bridge between Earth Engine + Python + maps
                #It helps convert EE images → visual layers
                #Very common in research + climate projects

